[feat/#35] 불필요 import 정리 및 파일 분리
